Make the password confirmation validation for has_secure_password optional (opt-out): 	* Wrap validates_confirmation_of call in option check 	* Rename ActiveModel::SecurePassword::InstanceMethodsOnActivation to 	  ActiveModel::SecurePassword::Authenticator for those who just want       the password= and authenticate methods 	* Update documentation 	* Consolidate test case models as mass_assignment_security_test
