--- activemodel/test/models/user.rb
@@ -1,8 +1,0 @@
-class User
-  include ActiveModel::Validations
-  include ActiveModel::SecurePassword
-
-  has_secure_password
-
-  attr_accessor :password_digest, :password_salt
-end
